+++ src/main/java/com/rakovets/course/java/core/practice/concurrency_thread_synchronization/producer_consumer/UserInputException.java
@@ -1,4 +1,4 @@
-package com.rakovets.course.java.core.practice.concurrency_thread_synchronization;
+package com.rakovets.course.java.core.practice.concurrency_thread_synchronization.producer_consumer;
 
 public class UserInputException extends RuntimeException{
     public UserInputException(String message) {
